+++ VTPG_Common.m
@@ -7,7 +7,7 @@
 //based off http://www.dribin.org/dave/blog/archives/2008/09/22/convert_to_nsstring/
 //
 static BOOL TypeCodeIsCharArray(const char *typeCode){
-	int lastCharOffset = strlen(typeCode) - 1;
+	int lastCharOffset = (int)(strlen(typeCode)) - 1;
 	int secondToLastCharOffset = lastCharOffset - 1 ;
 	
 	BOOL isCharArray = typeCode[0] == '[' &&
